fix:Phase-1 finish

+++ proj3/byow/Core/Engine.java
@@ -46,7 +46,7 @@ public class Engine {
         //
         // See proj3.byow.InputDemo for a demo of how you can make a nice clean interface
         // that works for many different input types.
-        int seed = Integer.parseInt(input.substring(1, input.length() - 1));
+        Long seed = Long.parseLong(input.substring(1, input.length() - 1));
         Random r = new Random(seed);
         TETile[][] finalWorldFrame = RandomCityGenerator.RandomGenerator(r);
         return finalWorldFrame;
